// cvat-ui/ltts_src/reducers/interfaces.ts
export interface UserState {
    users: any;
    count: number;
    fetching: boolean;
}